cppcheck: fixes
CBIORecv/Send are only assigned once.
IOCB_ReadCtx/WriteCtx only assigned once.
BuildMessage checks output and input wiht sizeOnly every call - state
machine doesn't cache these.
Renamed alias_tbl variables to something unique.
Other cppcheck fixes.
Also reworked pem_read_bio_key().

+++ IDE/IAR-MSP430/main.c
@@ -304,7 +304,7 @@ only_rng:
 
     printf(
         "ChaCha20/Poly1305 Encryption Start, 1000 itterations, %d bytes\r\n",
-        strlen((const char*)plaintext));
+        (int)strlen((const char*)plaintext));
     start = seconds;
     for (int i=0; i <= 1000; i++) {
         ret = wc_ChaCha20Poly1305_Encrypt(key, iv, aad, sizeof(aad), plaintext,
